Add generic http transport config
