+++ transporters/src/main/java/com/nextdoor/bender/ipc/generic/BenderHttpClientBuilder.java
@@ -1,0 +1,93 @@
+/*
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except
+ * in compliance with the License. You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under the License
+ * is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
+ * or implied. See the License for the specific language governing permissions and limitations under
+ * the License.
+ *
+ * Copyright 2017 Nextdoor.com, Inc
+ */
+
+package com.nextdoor.bender.ipc.generic;
+
+import java.security.KeyManagementException;
+import java.security.NoSuchAlgorithmException;
+import java.security.cert.CertificateException;
+import java.security.cert.X509Certificate;
+
+import javax.net.ssl.HostnameVerifier;
+import javax.net.ssl.SSLContext;
+import javax.net.ssl.SSLSession;
+import javax.net.ssl.TrustManager;
+import javax.net.ssl.X509TrustManager;
+
+import org.apache.http.impl.client.HttpClientBuilder;
+
+import com.nextdoor.bender.ipc.TransportFactoryInitException;
+
+public class BenderHttpClientBuilder extends HttpClientBuilder {
+  /**
+   * There isn't an easy way in java to trust non-self signed certs. Just allow all until java
+   * KeyStore functionality is added to Bender.
+   *
+   * @return a context that trusts all SSL certs
+   */
+  public static SSLContext getSSLContext() {
+    /*
+     * Create SSLContext and TrustManager that will trust all SSL certs.
+     *
+     * Copy pasta from http://stackoverflow.com/a/4837230
+     */
+    TrustManager tm = new X509TrustManager() {
+      public void checkClientTrusted(X509Certificate[] chain, String authType)
+          throws CertificateException {}
+
+      public void checkServerTrusted(X509Certificate[] chain, String authType)
+          throws CertificateException {}
+
+      public X509Certificate[] getAcceptedIssuers() {
+        return null;
+      }
+    };
+
+    SSLContext ctx;
+    try {
+      ctx = SSLContext.getInstance("TLS");
+    } catch (NoSuchAlgorithmException e) {
+      throw new TransportFactoryInitException("JVM does not have proper libraries for TSL");
+    }
+
+    try {
+      ctx.init(null, new TrustManager[] {tm}, new java.security.SecureRandom());
+    } catch (KeyManagementException e) {
+      throw new TransportFactoryInitException("Unable to init SSLContext with TrustManager", e);
+    }
+    return ctx;
+  }
+
+  public final HttpClientBuilder withSSL() {
+    /*
+     * All trusting SSL context
+     */
+    try {
+      this.setSSLContext(getSSLContext());
+    } catch (Exception e) {
+      throw new RuntimeException(e);
+    }
+
+    /*
+     * All trusting hostname verifier
+     */
+    this.setSSLHostnameVerifier(new HostnameVerifier() {
+      public boolean verify(String s, SSLSession sslSession) {
+        return true;
+      }
+    });
+
+    return this;
+  }
+}
