+++ transporters/src/main/java/com/nextdoor/bender/ipc/generic/GenericHttpTransport.java
@@ -19,6 +19,7 @@ import java.io.BufferedReader;
 import java.io.IOException;
 import java.io.InputStreamReader;
 import java.time.temporal.ChronoUnit;
+import java.util.LinkedHashMap;
 import java.util.concurrent.Callable;
 import java.util.zip.GZIPInputStream;
 
@@ -34,11 +35,13 @@ import org.apache.http.message.BasicHeader;
 import org.apache.http.util.EntityUtils;
 import org.apache.log4j.Logger;
 
+import com.amazonaws.services.lambda.runtime.Context;
 import com.evanlennick.retry4j.CallExecutor;
 import com.evanlennick.retry4j.RetryConfig;
 import com.evanlennick.retry4j.RetryConfigBuilder;
 import com.evanlennick.retry4j.exception.RetriesExhaustedException;
 import com.evanlennick.retry4j.exception.UnexpectedException;
+import com.nextdoor.bender.ipc.PartitionedTransport;
 import com.nextdoor.bender.ipc.TransportBuffer;
 import com.nextdoor.bender.ipc.TransportException;
 import com.nextdoor.bender.ipc.UnpartitionedTransport;
