+++ transporters/src/main/java/com/nextdoor/bender/ipc/scalyr/ScalyrTransportConfig.java
@@ -18,31 +18,21 @@ package com.nextdoor.bender.ipc.scalyr;
 
 import java.io.UnsupportedEncodingException;
 
-import javax.validation.constraints.Max;
-import javax.validation.constraints.Min;
-
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonTypeName;
 import com.kjetland.jackson.jsonSchema.annotations.JsonSchemaDefault;
 import com.kjetland.jackson.jsonSchema.annotations.JsonSchemaDescription;
-import com.nextdoor.bender.ipc.TransportConfig;
+import com.nextdoor.bender.ipc.generic.GenericHttpTransportConfig;
 import com.nextdoor.bender.utils.Passwords;
 
 @JsonTypeName("Scalyr")
 @JsonSchemaDescription("Writes events to a Scalyr endpoint.")
-public class ScalyrTransportConfig extends TransportConfig {
+public class ScalyrTransportConfig extends GenericHttpTransportConfig {
 
   @JsonSchemaDescription("Scalyr HTTP endpoint hostname.")
   @JsonSchemaDefault(value = "www.scalyr.com")
   @JsonProperty(required = false)
   private String hostname = "www.scalyr.com";
-
-  @JsonSchemaDescription("Scalyr HTTP endpoint port.")
-  @JsonSchemaDefault(value = "443")
-  @JsonProperty(required = false)
-  @Min(1)
-  @Max(65535)
-  private Integer port = 443;
 
   @JsonSchemaDescription("Scalyr auth token. If value is kms encrypted prefix with 'KMS='.")
   @JsonProperty(required = true)
@@ -53,57 +43,12 @@ public class ScalyrTransportConfig extends TransportConfig {
   @JsonProperty(required = false)
   private String parser = "json";
 
-  @JsonSchemaDescription("Use SSL connections (certificates are not validated).")
-  @JsonSchemaDefault(value = "true")
-  @JsonProperty(required = false)
-  private Boolean useSSL = true;
-
-  @JsonSchemaDescription("Use GZIP compression on HTTP calls.")
-  @JsonSchemaDefault(value = "false")
-  @JsonProperty(required = false)
-  private Boolean useGzip = false;
-
-  @JsonSchemaDescription("Maximum number of documents in bulk api call.")
-  @JsonSchemaDefault(value = "500")
-  @JsonProperty(required = false)
-  @Min(500)
-  @Max(100000)
-  private Integer batchSize = 500;
-
-  @JsonSchemaDescription("Number of retries to make when a put failure occurs.")
-  @JsonSchemaDefault(value = "0")
-  @JsonProperty(required = false)
-  @Min(0)
-  @Max(10)
-  private Integer retryCount = 0;
-
-  @JsonSchemaDescription("Initial delay between retries. If more than one retries specified exponential backoff is used.")
-  @JsonSchemaDefault(value = "1000")
-  @JsonProperty(required = false)
-  @Min(1)
-  @Max(60000)
-  private Long retryDelay = 1000l;
-
-  @JsonSchemaDescription("Socket timeout on HTTP connection.")
-  @JsonSchemaDefault(value = "40000")
-  @JsonProperty(required = false)
-  @Min(1000)
-  @Max(300000)
-  private Integer timeout = 40000;
-
   public String getHostname() {
     return hostname;
   }
 
   public void setHostname(String hostname) {
     this.hostname = hostname;
-  }
-
-  public Integer getPort() { return port;
-  }
-
-  public void setPort(Integer port) {
-    this.port = port;
   }
 
   public void setToken(String token) {
@@ -130,54 +75,6 @@ public class ScalyrTransportConfig extends TransportConfig {
 
   public void setParser(String parser) {
     this.parser = parser;
-  }
-
-  public Boolean isUseSSL() {
-    return useSSL;
-  }
-
-  public void setUseSSL(Boolean useSSL) {
-    this.useSSL = useSSL;
-  }
-
-  public Boolean isUseGzip() {
-    return useGzip;
-  }
-
-  public void setUseGzip(Boolean useGzip) {
-    this.useGzip = useGzip;
-  }
-
-  public Integer getBatchSize() {
-    return batchSize;
-  }
-
-  public void setBatchSize(Integer batchSize) {
-    this.batchSize = batchSize;
-  }
-
-  public Integer getRetryCount() {
-    return retryCount;
-  }
-
-  public void setRetryCount(Integer retryCount) {
-    this.retryCount = retryCount;
-  }
-
-  public Long getRetryDelay() {
-    return retryDelay;
-  }
-
-  public void setRetryDelay(Long retryDelay) {
-    this.retryDelay = retryDelay;
-  }
-
-  public Integer getTimeout() {
-    return timeout;
-  }
-
-  public void setTimeout(Integer timeout) {
-    this.timeout = timeout;
   }
 
   @Override
