+++ transporters/src/main/java/com/nextdoor/bender/ipc/scalyr/ScalyrTransportFactory.java
@@ -1,15 +1,13 @@
 /*
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except
+ * in compliance with the License. You may obtain a copy of the License at
  *
  * http://www.apache.org/licenses/LICENSE-2.0
  *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
+ * Unless required by applicable law or agreed to in writing, software distributed under the License
+ * is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
+ * or implied. See the License for the specific language governing permissions and limitations under
+ * the License.
  *
  * Copyright 2017 Nextdoor.com, Inc
  */
@@ -17,20 +15,10 @@
 package com.nextdoor.bender.ipc.scalyr;
 
 import java.io.IOException;
-import java.security.KeyManagementException;
-import java.security.NoSuchAlgorithmException;
-import java.security.cert.CertificateException;
-import java.security.cert.X509Certificate;
-
-import javax.net.ssl.SSLContext;
-import javax.net.ssl.TrustManager;
-import javax.net.ssl.X509TrustManager;
 
 import org.apache.http.config.SocketConfig;
-import org.apache.http.conn.ssl.SSLConnectionSocketFactory;
 import org.apache.http.impl.client.CloseableHttpClient;
 import org.apache.http.impl.client.HttpClientBuilder;
-import org.apache.http.impl.client.HttpClients;
 
 import com.nextdoor.bender.config.AbstractConfig;
 import com.nextdoor.bender.ipc.TransportBuffer;
@@ -38,6 +26,7 @@ import com.nextdoor.bender.ipc.TransportException;
 import com.nextdoor.bender.ipc.TransportFactory;
 import com.nextdoor.bender.ipc.TransportFactoryInitException;
 import com.nextdoor.bender.ipc.UnpartitionedTransport;
+import com.nextdoor.bender.ipc.generic.BenderHttpClientBuilder;
 import com.nextdoor.bender.ipc.generic.GenericHttpTransport;
 import com.nextdoor.bender.ipc.generic.GenericTransportBuffer;
 
@@ -61,8 +50,8 @@ public class ScalyrTransportFactory implements TransportFactory {
 
   @Override
   public UnpartitionedTransport newInstance() throws TransportFactoryInitException {
-    return new GenericHttpTransport(this.client, this.url,
-        this.config.isUseGzip(), this.config.getRetryCount(), this.config.getRetryDelay());
+    return new GenericHttpTransport(this.client, this.url, this.config.isUseGzip(),
+        this.config.getRetryCount(), this.config.getRetryDelay());
   }
 
   @Override
@@ -75,54 +64,13 @@ public class ScalyrTransportFactory implements TransportFactory {
     }
   }
 
-  /**
-   * There isn't an easy way in java to trust non-self signed certs. Just allow all until java
-   * KeyStore functionality is added to Bender.
-   *
-   * @return a context that trusts all SSL certs
-   */
-  private SSLContext getSSLContext() {
-    /*
-     * Create SSLContext and TrustManager that will trust all SSL certs.
-     *
-     * Copy pasta from http://stackoverflow.com/a/4837230
-     */
-    TrustManager tm = new X509TrustManager() {
-      public void checkClientTrusted(X509Certificate[] chain, String authType)
-          throws CertificateException {}
 
-      public void checkServerTrusted(X509Certificate[] chain, String authType)
-          throws CertificateException {}
-
-      public X509Certificate[] getAcceptedIssuers() {
-        return null;
-      }
-    };
-
-    SSLContext ctx;
-    try {
-      ctx = SSLContext.getInstance("TLS");
-    } catch (NoSuchAlgorithmException e) {
-      throw new TransportFactoryInitException("JVM does not have proper libraries for TSL");
-    }
-
-    try {
-      ctx.init(null, new TrustManager[] {tm}, new java.security.SecureRandom());
-    } catch (KeyManagementException e) {
-      throw new TransportFactoryInitException("Unable to init SSLContext with TrustManager", e);
-    }
-    return ctx;
-  }
 
   private CloseableHttpClient getHttpClient() throws TransportFactoryInitException {
-    HttpClientBuilder cb = HttpClients.custom();
+    HttpClientBuilder cb = BenderHttpClientBuilder.create();
 
     if (this.config.isUseSSL()) {
-      try {
-        cb.setSslcontext(getSSLContext());
-        cb = cb.setHostnameVerifier(SSLConnectionSocketFactory.ALLOW_ALL_HOSTNAME_VERIFIER);
-      } catch (Exception e) {
-      }
+      ((BenderHttpClientBuilder) (cb)).withSSL();
     }
 
     cb.setMaxConnTotal(this.config.getThreads());
@@ -152,10 +100,8 @@ public class ScalyrTransportFactory implements TransportFactory {
       confUrl += "http://";
     }
 
-    confUrl += this.config.getHostname() + ":" +
-               this.config.getPort() +
-               "/api/uploadLogs?parser=" + this.config.getParser() +
-               "&token=" + this.config.getToken();
+    confUrl += this.config.getHostname() + ":" + this.config.getPort() + "/api/uploadLogs?parser="
+        + this.config.getParser() + "&token=" + this.config.getToken();
 
     this.url = confUrl;
   }
